only allow substr and stripPrefix for permissions
