+++ backend/src/ee/services/dynamic-secret/providers/aws-elasticache.ts
@@ -8,11 +8,11 @@ import {
   ModifyReplicationGroupCommand,
   ModifyUserGroupCommand
 } from "@aws-sdk/client-elasticache";
+import handlebars from "handlebars";
 import { customAlphabet } from "nanoid";
 import { z } from "zod";
 
 import { TDynamicSecrets } from "@app/db/schemas";
-import { createHandlebarsClient } from "@app/lib/template/handlebars-client";
 import { CustomAWSHasher } from "@app/lib/aws/hashing";
 import { crypto } from "@app/lib/crypto";
 import { BadRequestError } from "@app/lib/errors";
@@ -211,7 +211,7 @@ export const AwsElastiCacheDatabaseProvider = (): TDynamicProviderFns => {
     const leasePassword = generatePassword();
     const leaseExpiration = new Date(expireAt).toISOString();
 
-    const creationStatement = createHandlebarsClient().compile(providerInputs.creationStatement, { noEscape: true })({
+    const creationStatement = handlebars.compile(providerInputs.creationStatement, { noEscape: true })({
       username: leaseUsername,
       password: leasePassword,
       expiration: leaseExpiration
@@ -255,7 +255,7 @@ export const AwsElastiCacheDatabaseProvider = (): TDynamicProviderFns => {
   const revoke = async (inputs: unknown, entityId: string) => {
     const providerInputs = await validateProviderInputs(inputs);
 
-    const revokeStatement = createHandlebarsClient().compile(providerInputs.revocationStatement, { noEscape: true })({
+    const revokeStatement = handlebars.compile(providerInputs.revocationStatement, { noEscape: true })({
       username: entityId
     });
     const parsedStatement = DeleteElasticCacheUserSchema.parse(JSON.parse(revokeStatement));
