+++ backend/src/ee/services/dynamic-secret/providers/azure-sql-database.ts
@@ -1,8 +1,7 @@
+import handlebars from "handlebars";
 import knex from "knex";
 import RE2 from "re2";
 import { z } from "zod";
-
-import { createHandlebarsClient } from "@app/lib/template/handlebars-client";
 
 import { TDynamicSecrets } from "@app/db/schemas";
 import { crypto } from "@app/lib/crypto/cryptography";
@@ -303,7 +302,7 @@ export const AzureSqlDatabaseProvider = ({
       );
 
       try {
-        const masterCreationStatement = createHandlebarsClient().compile(providerInputs.masterCreationStatement, { noEscape: true })({
+        const masterCreationStatement = handlebars.compile(providerInputs.masterCreationStatement, { noEscape: true })({
           username,
           password,
           expiration,
@@ -337,7 +336,7 @@ export const AzureSqlDatabaseProvider = ({
       });
 
       try {
-        const creationStatement = createHandlebarsClient().compile(providerInputs.creationStatement, { noEscape: true })({
+        const creationStatement = handlebars.compile(providerInputs.creationStatement, { noEscape: true })({
           username,
           password,
           expiration,
@@ -378,7 +377,7 @@ export const AzureSqlDatabaseProvider = ({
     const { database, masterDatabase } = providerInputs;
 
     const gatewayCallback = async (host = providerInputs.host, port = providerInputs.port) => {
-      const revokeStatement = createHandlebarsClient().compile(providerInputs.revocationStatement, { noEscape: true })({
+      const revokeStatement = handlebars.compile(providerInputs.revocationStatement, { noEscape: true })({
         username,
         database
       });
@@ -501,7 +500,7 @@ export const AzureSqlDatabaseProvider = ({
       const expiration = new Date(expireAt).toISOString();
       const { database } = providerInputs;
 
-      const renewStatement = createHandlebarsClient().compile(providerInputs.renewStatement, { noEscape: true })({
+      const renewStatement = handlebars.compile(providerInputs.renewStatement, { noEscape: true })({
         username: entityId,
         expiration,
         database
