+++ backend/src/ee/services/dynamic-secret/providers/sap-ase.ts
@@ -1,8 +1,7 @@
+import handlebars from "handlebars";
 import { customAlphabet } from "nanoid";
 import odbc from "odbc";
 import { z } from "zod";
-
-import { createHandlebarsClient } from "@app/lib/template/handlebars-client";
 
 import { TDynamicSecrets } from "@app/db/schemas";
 import { BadRequestError } from "@app/lib/errors";
@@ -117,7 +116,7 @@ export const SapAseProvider = (): TDynamicProviderFns => {
     const client = await $getClient(providerInputs);
     const masterClient = await $getClient(providerInputs, true);
 
-    const creationStatement = createHandlebarsClient().compile(providerInputs.creationStatement, { noEscape: true })({
+    const creationStatement = handlebars.compile(providerInputs.creationStatement, { noEscape: true })({
       username,
       password
     });
@@ -149,7 +148,7 @@ export const SapAseProvider = (): TDynamicProviderFns => {
   const revoke = async (inputs: unknown, username: string) => {
     const providerInputs = await validateProviderInputs(inputs);
 
-    const revokeStatement = createHandlebarsClient().compile(providerInputs.revocationStatement, { noEscape: true })({
+    const revokeStatement = handlebars.compile(providerInputs.revocationStatement, { noEscape: true })({
       username
     });
