+++ backend/src/ee/services/dynamic-secret/providers/snowflake.ts
@@ -1,8 +1,7 @@
+import handlebars from "handlebars";
 import { customAlphabet } from "nanoid";
 import snowflake from "snowflake-sdk";
 import { z } from "zod";
-
-import { createHandlebarsClient } from "@app/lib/template/handlebars-client";
 
 import { TDynamicSecrets } from "@app/db/schemas";
 import { BadRequestError } from "@app/lib/errors";
@@ -110,7 +109,7 @@ export const SnowflakeProvider = (): TDynamicProviderFns => {
 
     try {
       const expiration = getDaysToExpiry(new Date(expireAt));
-      const creationStatement = createHandlebarsClient().compile(providerInputs.creationStatement, { noEscape: true })({
+      const creationStatement = handlebars.compile(providerInputs.creationStatement, { noEscape: true })({
         username,
         password,
         expiration
@@ -147,7 +146,7 @@ export const SnowflakeProvider = (): TDynamicProviderFns => {
     const client = await $getClient(providerInputs);
 
     try {
-      const revokeStatement = createHandlebarsClient().compile(providerInputs.revocationStatement, { noEscape: true })({ username });
+      const revokeStatement = handlebars.compile(providerInputs.revocationStatement, { noEscape: true })({ username });
 
       await new Promise((resolve, reject) => {
         client.execute({
@@ -182,7 +181,7 @@ export const SnowflakeProvider = (): TDynamicProviderFns => {
 
     try {
       const expiration = getDaysToExpiry(new Date(expireAt));
-      const renewStatement = createHandlebarsClient().compile(providerInputs.renewStatement, { noEscape: true })({
+      const renewStatement = handlebars.compile(providerInputs.renewStatement, { noEscape: true })({
         username: entityId,
         expiration
       });
