+++ backend/src/ee/services/dynamic-secret/providers/sql-database.ts
@@ -1,8 +1,7 @@
+import handlebars from "handlebars";
 import knex from "knex";
 import RE2 from "re2";
 import { z } from "zod";
-
-import { createHandlebarsClient } from "@app/lib/template/handlebars-client";
 
 import { TDynamicSecrets } from "@app/db/schemas";
 import { crypto } from "@app/lib/crypto/cryptography";
@@ -322,7 +321,7 @@ export const SqlDatabaseProvider = ({
       try {
         const expiration = new Date(expireAt).toISOString();
 
-        const creationStatement = createHandlebarsClient().compile(providerInputs.creationStatement, { noEscape: true })({
+        const creationStatement = handlebars.compile(providerInputs.creationStatement, { noEscape: true })({
           username,
           password,
           expiration,
@@ -368,7 +367,7 @@ export const SqlDatabaseProvider = ({
         originalHost: providerInputs.host
       });
       try {
-        const revokeStatement = createHandlebarsClient().compile(providerInputs.revocationStatement, { noEscape: true })({
+        const revokeStatement = handlebars.compile(providerInputs.revocationStatement, { noEscape: true })({
           username,
           database
         });
@@ -413,7 +412,7 @@ export const SqlDatabaseProvider = ({
       const expiration = new Date(expireAt).toISOString();
       const { database } = providerInputs;
 
-      const renewStatement = createHandlebarsClient().compile(providerInputs.renewStatement, { noEscape: true })({
+      const renewStatement = handlebars.compile(providerInputs.renewStatement, { noEscape: true })({
         username: entityId,
         expiration,
         database
